Add solution for problem `B - Count Subrectangles` Link: https://codeforces.com/contest/1323/problem/B Tags: binary search, greedy, implementation, *1500 Ref: https://codeforces.com/contest/1323/submission/72733632

+++ codeforces/1323/B.cpp
@@ -1,8 +1,8 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int c(vector<int> & v, int x) {
-    int res = 0;
+long long c(vector<int> & v, int x) {
+    long long res = 0;
     for(int & i : v) {
         if(i >= x)
             res += i - x + 1;
